+++ changelog.md
@@ -5,3 +5,4 @@
 #### 1.1.1
 - UI improvements (config selector visible only when multiple configs are available)
 - Comment improvements 
+- Fixed comparing local and remote versions